@@ -31,7 +31,7 @@ install_requires =
     napari
     numpy
     tinycss2
-python_requires = >=3.8
+python_requires = >=3.9
 include_package_data = True
 package_dir =
     =src
